Add GameStateManager
