+++ Assets/Scripts/GameStateManager.cs.meta
@@ -1,10 +1,10 @@
 fileFormatVersion: 2
-guid: b4ef247928707426e8ec2aae6eee4c5d
+guid: 513a786bff8474b45a37166c87e32d17
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
-  executionOrder: 800
+  executionOrder: 700
   icon: {instanceID: 0}
   userData: 
   assetBundleName:
