+++ Assets/Scripts/PlayerKeyboardManager.cs
@@ -14,19 +14,14 @@ public class PlayerKeyboardManager : MonoBehaviour
     private bool crouchFlag;
     public bool virtualKeyboard;
 
-    private int gamemode; //{1 = segue, 2 = conversing, 3 = exploring, 4 = death, 5 = trading}
-    private bool inventoryInScene = true; //Set to false if no inventory in scene (Ex. QuakeHouse)
-    private bool npcInteractedInScene = true;
-    private bool pointsInScene = true;
+    private Gamemode currentGamemode;
 
     private ReferenceManager referenceManager;
     private TradeManagerUI tradeManager;
     private DialogueManager dialogueManager;
     private TooltipManager tooltipManager;
-    private GameObject deathCanvas;
-    private GameObject segueCanvas;
+    private GameStateManager gameStateManager;
     private GameObject npcInteractedCanvas;
-    private GameObject metersCanvas;
 
     public bool leftTrading = false;
 
@@ -61,9 +56,9 @@ public class PlayerKeyboardManager : MonoBehaviour
         tradeManager = referenceManager.tradeCanvas.GetComponent<TradeManagerUI>();
         dialogueManager = referenceManager.dialogueCanvas.GetComponent<DialogueManager>();
         tooltipManager = referenceManager.tooltipManager;
+        gameStateManager = referenceManager.gameStateManager.GetComponent<GameStateManager>();
         inventory = referenceManager.inventoryCanvas.GetComponent<Inventory>();
         inventoryUI = referenceManager.inventoryCanvas.GetComponent<InventoryUI>();
-        metersCanvas = referenceManager.metersCanvas;
         npcInteractedCanvas = referenceManager.npcInteractedCanvas;
     }
 
@@ -74,9 +69,6 @@ public class PlayerKeyboardManager : MonoBehaviour
         tooltipManager.HandleTooltip();
 
         player = referenceManager.player.GetComponent<PlayerMover>();
-        deathCanvas = referenceManager.deathCanvas;
-        deathCanvas.SetActive(false);
-        segueCanvas = referenceManager.segueCanvas;
         cursorLocation = 0;
 
         unselected = Resources.Load<Sprite>("UnselectedSlot 1");
@@ -91,78 +83,17 @@ public class PlayerKeyboardManager : MonoBehaviour
                 index++;
             }
         }
+        
+        if (gameStateManager.NPCInteractedInScene())
+        {
+            referenceManager.npcInteractedCanvas.SetActive(true);
 
-        //Handle start of scene things
-        if (SceneManager.GetActiveScene().name.Contains("Quake"))
-        {
-            GlobalControls.globalControlsProperties.Remove("metersEnabled");
-        }
-        else
-        {
-            GlobalControls.globalControlsProperties.Add("metersEnabled");
-        }
-
-        if (SceneManager.GetActiveScene().name.Equals("GameEnd"))
-        {
-            inventoryInScene = false;
-            GlobalControls.globalControlsProperties.Remove("objectivesEnabled");
-            npcInteractedInScene = false;
-            SetExploring();
-        }
-        else if (SceneManager.GetActiveScene().name.Equals("QuakeApartment") ||
-                 SceneManager.GetActiveScene().name.Equals("QuakeHouse"))
-        {
-            if (!inventory.gameObject.activeSelf) inventory.gameObject.SetActive(true);
-            inventory.SetAvailableSlots(2);
-            pointsInScene = false;
-            npcInteractedInScene = false;
-            GlobalControls.CurrentObjective = 3;
-            SetSegue();
-        }
-        else if (SceneManager.GetActiveScene().name.Equals("PreQuakeHouse") ||
-                 SceneManager.GetActiveScene().name.Equals("PreQuakeApartment"))
-        {
-            if (!inventory.gameObject.activeSelf) inventory.gameObject.SetActive(true);
-            inventory.SetAvailableSlots(1);
-            npcInteractedInScene = false;
-            pointsInScene = false;
-            if (GlobalControls.globalControlsProperties.Contains("apartmentCondition"))
-                GlobalControls.CurrentObjective = 2;
-            else if (!GlobalControls.globalControlsProperties.Contains("apartmentCondition"))
-                GlobalControls.CurrentObjective = 1;
-            SetExploring();
-        }
-        else if (SceneManager.GetActiveScene().name.Equals("StrategicMap"))
-        {
-            gameObject.GetComponent<StrategicMapKeyboardController>().enabled = true;
-            enabled = false;
-        }
-        else if (GlobalControls.CurrentObjective <= 4)
-        {
-            GlobalControls.CurrentObjective = 5;
-            SetExploring();
-        }
-        else SetExploring();
-
-        if (!SceneManager.GetActiveScene().name.Equals("StrategicMap") &&
-            GlobalControls.globalControlsProperties.Contains("metersEnabled"))
-        {
-            if (!GlobalControls.globalControlsProperties.Contains("poopTaskCompleted") &&
-                GlobalControls.PoopTimeLeft == 0)
+            for (int i = 0; i < validNPCInputs.Length; i++)
             {
-                Debug.Log("You died by poop meter going to zero!!");
-                GameObject.Find("Managers").GetComponent<ReferenceManager>().deathManager.GetComponent<PlayerDeath>()
-                    .KillPlayer(metersCanvas, 4);
-            }
-
-            if (!GlobalControls.globalControlsProperties.Contains("waterTaskCompleted") &&
-                GlobalControls.WaterTimeLeft == 0)
-            {
-                Debug.Log("You died of thirst!");
-                GameObject.Find("Managers").GetComponent<ReferenceManager>().deathManager.GetComponent<PlayerDeath>()
-                    .KillPlayer(metersCanvas, 2);
+                npcFrames[i].GetComponent<Image>().sprite = unselected;
             }
         }
+        else referenceManager.npcInteractedCanvas.SetActive(false);
     }
 
     /// <summary>
@@ -177,6 +108,7 @@ public class PlayerKeyboardManager : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
+        currentGamemode = gameStateManager.currentGamemode;
         if (!virtualKeyboard)
         {
             if (Input.GetKeyDown(KeyCode.C)) keyDown = KeyCode.C;
@@ -187,22 +119,22 @@ public class PlayerKeyboardManager : MonoBehaviour
             else if (Input.GetKeyDown(KeyCode.Period)) keyDown = KeyCode.Period;
             else if (Input.GetKeyDown(KeyCode.LeftBracket)) keyDown = KeyCode.LeftBracket;
             else if (Input.GetKeyDown(KeyCode.RightBracket)) keyDown = KeyCode.RightBracket;
-            else if (gamemode != 2 && Input.GetKey(KeyCode.W)) keyDown = KeyCode.W;
-            else if (gamemode != 2 && Input.GetKey(KeyCode.A)) keyDown = KeyCode.A;
-            else if (gamemode != 2 && Input.GetKey(KeyCode.S)) keyDown = KeyCode.S;
-            else if (gamemode != 2 && Input.GetKey(KeyCode.D)) keyDown = KeyCode.D;
-            else if (gamemode == 2 && Input.GetKeyDown(KeyCode.W)) keyDown = KeyCode.W;
-            else if (gamemode == 2 && Input.GetKeyDown(KeyCode.A)) keyDown = KeyCode.A;
-            else if (gamemode == 2 && Input.GetKeyDown(KeyCode.S)) keyDown = KeyCode.S;
-            else if (gamemode == 2 && Input.GetKeyDown(KeyCode.D)) keyDown = KeyCode.D;
-            else if (gamemode != 2 && Input.GetKey(KeyCode.UpArrow)) keyDown = KeyCode.W;
-            else if (gamemode != 2 && Input.GetKey(KeyCode.LeftArrow)) keyDown = KeyCode.A;
-            else if (gamemode != 2 && Input.GetKey(KeyCode.DownArrow)) keyDown = KeyCode.S;
-            else if (gamemode != 2 && Input.GetKey(KeyCode.RightArrow)) keyDown = KeyCode.D;
-            else if (gamemode == 2 && Input.GetKeyDown(KeyCode.UpArrow)) keyDown = KeyCode.W;
-            else if (gamemode == 2 && Input.GetKeyDown(KeyCode.LeftArrow)) keyDown = KeyCode.A;
-            else if (gamemode == 2 && Input.GetKeyDown(KeyCode.DownArrow)) keyDown = KeyCode.S;
-            else if (gamemode == 2 && Input.GetKeyDown(KeyCode.RightArrow)) keyDown = KeyCode.D;
+            else if (currentGamemode != Gamemode.Conversing && Input.GetKey(KeyCode.W)) keyDown = KeyCode.W;
+            else if (currentGamemode != Gamemode.Conversing && Input.GetKey(KeyCode.A)) keyDown = KeyCode.A;
+            else if (currentGamemode != Gamemode.Conversing && Input.GetKey(KeyCode.S)) keyDown = KeyCode.S;
+            else if (currentGamemode != Gamemode.Conversing && Input.GetKey(KeyCode.D)) keyDown = KeyCode.D;
+            else if (currentGamemode == Gamemode.Conversing && Input.GetKeyDown(KeyCode.W)) keyDown = KeyCode.W;
+            else if (currentGamemode == Gamemode.Conversing && Input.GetKeyDown(KeyCode.A)) keyDown = KeyCode.A;
+            else if (currentGamemode == Gamemode.Conversing && Input.GetKeyDown(KeyCode.S)) keyDown = KeyCode.S;
+            else if (currentGamemode == Gamemode.Conversing && Input.GetKeyDown(KeyCode.D)) keyDown = KeyCode.D;
+            else if (currentGamemode != Gamemode.Conversing && Input.GetKey(KeyCode.UpArrow)) keyDown = KeyCode.W;
+            else if (currentGamemode != Gamemode.Conversing && Input.GetKey(KeyCode.LeftArrow)) keyDown = KeyCode.A;
+            else if (currentGamemode != Gamemode.Conversing && Input.GetKey(KeyCode.DownArrow)) keyDown = KeyCode.S;
+            else if (currentGamemode != Gamemode.Conversing && Input.GetKey(KeyCode.RightArrow)) keyDown = KeyCode.D;
+            else if (currentGamemode == Gamemode.Conversing && Input.GetKeyDown(KeyCode.UpArrow)) keyDown = KeyCode.W;
+            else if (currentGamemode == Gamemode.Conversing && Input.GetKeyDown(KeyCode.LeftArrow)) keyDown = KeyCode.A;
+            else if (currentGamemode == Gamemode.Conversing && Input.GetKeyDown(KeyCode.DownArrow)) keyDown = KeyCode.S;
+            else if (currentGamemode == Gamemode.Conversing && Input.GetKeyDown(KeyCode.RightArrow)) keyDown = KeyCode.D;
             foreach (var key in validInputs)
             {
                 if (Input.GetKeyDown(key))
@@ -213,23 +145,23 @@ public class PlayerKeyboardManager : MonoBehaviour
             }
         }
 
-        if (gamemode == 1)
+        if (currentGamemode == Gamemode.Segue)
         {
             UpdateSegue();
         }
-        else if (gamemode == 2) //if talking to an npc
+        else if (currentGamemode == Gamemode.Conversing) //if talking to an npc
         {
             UpdateConversing();
         }
-        else if (gamemode == 3) //if moving around the map
+        else if (currentGamemode == Gamemode.Exploring) //if moving around the map
         {
             UpdateExploring();
         }
-        else if (gamemode == 4) //if ded
+        else if (currentGamemode == Gamemode.Death) //if ded
         {
             UpdateDeath();
         }
-        else if (gamemode == 5) //if trading with an npc
+        else if (currentGamemode == Gamemode.Trading) //if trading with an npc
         {
             UpdateTrading();
         }
@@ -241,7 +173,7 @@ public class PlayerKeyboardManager : MonoBehaviour
     {
         if (keyDown.Equals(KeyCode.Space))
         {
-            SetExploring();
+            gameStateManager.SetExploring();
         }
     }
 
@@ -305,7 +237,7 @@ public class PlayerKeyboardManager : MonoBehaviour
 
         if (keyDown.Equals(KeyCode.Escape))
         {
-            SetExploring();
+            gameStateManager.SetExploring();
         }
 
         //Selects an option from the player options
@@ -499,7 +431,7 @@ public class PlayerKeyboardManager : MonoBehaviour
         else if (cursorLocation >= inventory.items.Length && npcInteractedCanvas.activeSelf)
             SelectSlotsNPCInteracted();
 
-        if (inventoryInScene && cursorLocation < inventory.items.Length && keyDown.Equals(KeyCode.Space))
+        if (gameStateManager.InventoryInScene() && cursorLocation < inventory.items.Length && keyDown.Equals(KeyCode.Space))
             inventory.PickUpOrDrop();
 
         if (inventory && npcInteractedCanvas.activeSelf &&
@@ -626,214 +558,13 @@ public class PlayerKeyboardManager : MonoBehaviour
         }
     }
 
-    public void SetExploring()
+    public void SetCursorLocation(int location)
     {
-        gamemode = 3;
-
-        deathCanvas.SetActive(false);
-        segueCanvas.SetActive(false);
-        referenceManager.player.GetComponent<PlayerMover>().enabled = true;
-        if (GlobalControls.globalControlsProperties.Contains("metersEnabled"))
-            referenceManager.metersCanvas.SetActive(true);
-        else if (!GlobalControls.globalControlsProperties.Contains("metersEnabled"))
-            referenceManager.metersCanvas.SetActive(false);
-        referenceManager.dialogueCanvas.SetActive(false);
-        referenceManager.tradeCanvas.SetActive(false);
-        if (GlobalControls.globalControlsProperties.Contains("objectivesEnabled"))
-        {
-            tooltipManager.SetObjectivesActive();
-            referenceManager.objectiveManager.UpdateObjectiveBanner();
-        }
-        else if (!GlobalControls.globalControlsProperties.Contains("objectivesEnabled"))
-            tooltipManager.SetObjectivesInactive();
-
-
-        if (GlobalControls.globalControlsProperties.Contains("tooltipsEnabled"))
-        {
-            tooltipManager.SetTooltipsActive();
-        }
-        else if (!GlobalControls.globalControlsProperties.Contains("tooltipsEnabled"))
-            tooltipManager.SetTooltipsInactive();
-
-        if (inventoryInScene)
-        {
-            referenceManager.tooltipCanvas.SetActive(true);
-            referenceManager.inventoryCanvas.SetActive(true);
-
-            cursorLocation = 0;
-
-            //Update to show Inventory selected
-            inventoryUI.EnableSelectedSlot();
-            inventory.SelectSlotNumber(0);
-        }
-
-        if (pointsInScene)
-        {
-            referenceManager.pointsText.gameObject.SetActive(true);
-            referenceManager.pointsText.GetComponentInChildren<Text>().text = GlobalControls.CurrentPoints.ToString();
-        }
-        else if (!pointsInScene) referenceManager.pointsText.gameObject.SetActive(false);
-
-        if (npcInteractedInScene)
-        {
-            referenceManager.npcInteractedCanvas.SetActive(true);
-
-            for (int i = 0; i < validNPCInputs.Length; i++)
-            {
-                npcFrames[i].GetComponent<Image>().sprite = unselected;
-            }
-        }
-        else if (!npcInteractedInScene) referenceManager.npcInteractedCanvas.SetActive(false);
-
-        tooltipManager.SetNPCInventoryTooltipInactive();
-
-        if (GlobalControls.globalControlsProperties.Contains("keybindsEnabled"))
-        {
-            referenceManager.keybinds.SetActive(true);
-            string exploringText = GlobalControls.Keybinds["Exploring"];
-            if (!npcInteractedInScene) exploringText = exploringText.Replace("\n[ ] => Switch inventory", "");
-            if (inventoryInScene)
-            {
-                Image[] images = inventory.gameObject.GetComponentsInChildren<Image>(true);
-                foreach (Image image in images)
-                {
-                    if (image.gameObject.name.Equals("Frame 1") && !image.gameObject.activeSelf)
-                    {
-                        exploringText = exploringText.Replace("\n< > => Move slots", "");
-                        break;
-                    }
-                }
-            }
-
-            referenceManager.keybinds.GetComponentInChildren<Text>().text = exploringText;
-        }
-        else if (GlobalControls.globalControlsProperties.Contains("keybindsEnabled"))
-            referenceManager.keybinds.SetActive(false);
+        cursorLocation = location;
     }
-
-    public void SetConversing()
+    
+    public void SetInventoryNumber(int number)
     {
-        gamemode = 2;
-        cursorLocation = 0;
-
-        deathCanvas.SetActive(false);
-        segueCanvas.SetActive(false);
-        referenceManager.tooltipCanvas.SetActive(true);
-        referenceManager.player.GetComponent<PlayerMover>().enabled = false;
-        referenceManager.metersCanvas.SetActive(false);
-        referenceManager.inventoryCanvas.SetActive(false);
-        referenceManager.dialogueCanvas.SetActive(true);
-        referenceManager.tradeCanvas.SetActive(false);
-        referenceManager.npcInteractedCanvas.SetActive(false);
-        if (GlobalControls.globalControlsProperties.Contains("objectivesEnabled"))
-        {
-            tooltipManager.SetObjectivesActive();
-            referenceManager.objectiveManager.UpdateObjectiveBanner();
-        }
-        else if (!GlobalControls.globalControlsProperties.Contains("objectivesEnabled"))
-            tooltipManager.SetObjectivesInactive();
-
-        if (GlobalControls.globalControlsProperties.Contains("keybindsEnabled"))
-        {
-            referenceManager.keybinds.SetActive(true);
-            referenceManager.keybinds.GetComponentInChildren<Text>().text = GlobalControls.Keybinds["Conversing"];
-        }
-        else if (GlobalControls.globalControlsProperties.Contains("keybindsEnabled"))
-            referenceManager.keybinds.SetActive(false);
-
-        tooltipManager.SetTooltipsInactive();
-
-        referenceManager.dialogueCanvas.GetComponent<DialogueManager>().BeginConversation();
-        tooltipManager.SetNPCInventoryTooltipInactive();
-        if (pointsInScene)
-        {
-            referenceManager.pointsText.gameObject.SetActive(true);
-            referenceManager.pointsText.GetComponentInChildren<Text>().text = GlobalControls.CurrentPoints.ToString();
-        }
-        else if (!pointsInScene) referenceManager.pointsText.gameObject.SetActive(false);
-    }
-
-    public void SetTrading()
-    {
-        gamemode = 5;
-        cursorLocation = 0;
-        inventoryNumber = 0;
-
-        deathCanvas.SetActive(false);
-        segueCanvas.SetActive(false);
-        referenceManager.tooltipCanvas.SetActive(true);
-        referenceManager.player.GetComponent<PlayerMover>().enabled = false;
-        referenceManager.metersCanvas.SetActive(false);
-        referenceManager.inventoryCanvas.SetActive(false);
-        referenceManager.dialogueCanvas.SetActive(false);
-        referenceManager.tradeCanvas.SetActive(true);
-        referenceManager.npcInteractedCanvas.SetActive(false);
-        tooltipManager.SetObjectivesInactive();
-        if (GlobalControls.globalControlsProperties.Contains("tooltipsEnabled"))
-        {
-            tooltipManager.SetTooltipsActive();
-        }
-        else if (!GlobalControls.globalControlsProperties.Contains("tooltipsEnabled"))
-            tooltipManager.SetTooltipsInactive();
-
-        if (GlobalControls.globalControlsProperties.Contains("keybindsEnabled"))
-        {
-            referenceManager.keybinds.SetActive(true);
-            referenceManager.keybinds.GetComponentInChildren<Text>().text = GlobalControls.Keybinds["Trading"];
-        }
-        else if (GlobalControls.globalControlsProperties.Contains("keybindsEnabled"))
-            referenceManager.keybinds.SetActive(false);
-
-        if (GlobalControls.globalControlsProperties.Contains("objectivesEnabled"))
-        {
-            tooltipManager.SetObjectivesActive();
-            referenceManager.objectiveManager.UpdateObjectiveBanner();
-        }
-        else if (!GlobalControls.globalControlsProperties.Contains("objectivesEnabled"))
-            tooltipManager.SetObjectivesInactive();
-
-        if (pointsInScene)
-        {
-            referenceManager.pointsText.gameObject.SetActive(true);
-            referenceManager.pointsText.GetComponentInChildren<Text>().text = GlobalControls.CurrentPoints.ToString();
-        }
-        else if (!pointsInScene) referenceManager.pointsText.gameObject.SetActive(false);
-
-        referenceManager.tradeCanvas.GetComponent<TradeManagerUI>().BeginTrading();
-        tooltipManager.SetNPCInventoryTooltipInactive();
-    }
-
-    public void SetSegue()
-    {
-        gamemode = 1;
-
-        referenceManager.tooltipCanvas.SetActive(false);
-        referenceManager.player.GetComponent<PlayerMover>().enabled = false;
-        referenceManager.metersCanvas.SetActive(false);
-        referenceManager.inventoryCanvas.SetActive(false);
-        referenceManager.dialogueCanvas.SetActive(false);
-        referenceManager.tradeCanvas.SetActive(false);
-        referenceManager.npcInteractedCanvas.SetActive(false);
-
-        deathCanvas.SetActive(false);
-        segueCanvas.SetActive(true);
-        referenceManager.pointsText.gameObject.SetActive(false);
-    }
-
-    public void SetDeath()
-    {
-        gamemode = 4;
-
-        referenceManager.tooltipCanvas.SetActive(false);
-        referenceManager.player.GetComponent<PlayerMover>().enabled = false;
-        referenceManager.metersCanvas.SetActive(false);
-        referenceManager.inventoryCanvas.SetActive(false);
-        referenceManager.dialogueCanvas.SetActive(false);
-        referenceManager.tradeCanvas.SetActive(false);
-        referenceManager.npcInteractedCanvas.SetActive(false);
-
-        deathCanvas.SetActive(true);
-        segueCanvas.SetActive(false);
-        referenceManager.pointsText.gameObject.SetActive(false);
+        inventoryNumber = number;
     }
 }
