+++ Assets/Scripts/ReferenceManager.cs
@@ -27,6 +27,7 @@ public class ReferenceManager : MonoBehaviour
     public GameObject keybinds;
     public ObjectiveManager objectiveManager;
     public TooltipManager tooltipManager;
+    public GameObject gameStateManager;
     public GameObject pointsText;
 
     public GameObject itemLoader;
@@ -77,6 +78,7 @@ public class ReferenceManager : MonoBehaviour
         sceneManagement = GameObject.Find("Scene Management");
         itemLoader = GameObject.Find("Item Manager");
         objectiveManager = GameObject.Find("Objective Manager").GetComponent<ObjectiveManager>();
+        gameStateManager = GameObject.Find("Game State Manager");
         tooltipManager = tooltipCanvas.GetComponent<TooltipManager>();
         foreach (Transform child in tooltipCanvas.GetComponentsInChildren<Transform>(true))
         {
